flask_version/requirements.txt
google-cloud-texttospeech==2.15.0
Flask==2.2.5